advanced-multitoolV1.5/requirements.txt
requests
pystyle
datetime
selenium
deep_translator
colorama
phonenumbers
python-whois
geocoder
pycryptodome
discord.py
pywin32
tk
customtkinter
pyinstaller
psutil
pillow
bs4
lxml
requests==2.26.0
Pillow==8.4.0
fade
colorama